requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
pytz>=2023.3

# OpenAI API
openai>=1.3.0

# Gmail API
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0

# Additional utilities
lxml>=4.9.3
html5lib>=1.1
python-dotenv>=1.0.0
PyPDF2>=3.0.0

# JavaScript scraping
selenium>=4.15.0
webdriver-manager>=4.0.0
